chore: Move create policy endpoint to endpoints section of app
