+++ webapp/endpoints/models.py
@@ -97,6 +97,52 @@ def get_policies(store_id: str, model_name: str):
 
 
 @models.route(
+    "/api/store/<store_id>/models/<model_name>/policies", methods=["POST"]
+)
+@login_required
+@exchange_required
+def create_policy(store_id: str, model_name: str):
+    """
+    Creat policy for a store model.
+
+    Args:
+        store_id (str): The ID of the store.
+        model_name (str): The name of the model.
+
+    Returns:
+        dict: A dictionary containing the response message and success
+    """
+    signing_key = flask.request.form.get("signing_key")
+    res = {}
+    try:
+        signing_keys_data = publisher_gateway.get_store_signing_keys(
+            flask.session, store_id
+        )
+        signing_keys = [key["sha3-384"] for key in signing_keys_data]
+
+        if not signing_key:
+            res["message"] = "Signing key required"
+            res["success"] = False
+            return make_response(res, 500)
+
+        if signing_key in signing_keys:
+            publisher_gateway.create_store_model_policy(
+                flask.session, store_id, model_name, signing_key
+            )
+            res["success"] = True
+        else:
+            res["message"] = "Invalid signing key"
+            res["success"] = False
+    except StoreApiResponseErrorList as error_list:
+        res["success"] = False
+        res["message"] = error_list.errors[0]["message"]
+
+    if res["success"]:
+        return make_response(res, 200)
+    return make_response(res, 500)
+
+
+@models.route(
     "/api/store/<store_id>/models/<model_name>/policies/<revision>",
     methods=["DELETE"],
 )
